move resend_confirmation_email to a new mod
