+++ ft-email-auth-provider/src/handlers/mod.rs
@@ -1,3 +1,4 @@
 pub mod create_account;
 pub mod login;
 pub mod confirm_email;
+pub mod resend_confirmation_email;
